+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add `--hook` flag to `agent-guide` command to output JSON format for SessionStart hooks
+
 ## [3.8.0] - 2026-05-14
 
 ### Added